ashr segfault on travis?

+++ include/coreir/ir/dynamic_bit_vector.h
@@ -877,7 +877,7 @@ namespace bsim {
       res.set(i - shift_int, a.get(i));
     }
 
-    int last_index = (int)a.bitLength() - shift_int;
+    int last_index = (int)a.bitLength() - (int)shift_int;
     for (int i = a.bitLength() - 1; i >= last_index && i >= 0; i--) {
       res.set(i, sign_bit);
     }
